feat: update user-menu storybook
